add math, delta time

+++ GameClient/Source/Object/Object.h
@@ -6,7 +6,7 @@ class Object
 {
 public:
 	virtual void Start() = 0;
-	virtual void Update( float dt ) = 0;
+	virtual void Update( double dt ) = 0;
 	virtual void Draw() = 0;
 
 private:
