+++ GameClient/Source/Object/World.cpp
@@ -16,7 +16,7 @@ void World::Start()
 		}
 	}
 }
-void World::Update( float dt )
+void World::Update( double dt )
 {
 
 }
